move story assets under story folder
