+++ scripts/stories.cjs
@@ -3,7 +3,7 @@ const fg = require('fast-glob');
 const { rewritePaths } = require('typescript-rewrite-paths');
 const path = require('path');
 
-const EXCLUDED_PATHS = ['./examples'];
+const EXCLUDED_PATHS = ['./examples', './assets'];
 
 /**
  * Replace all the paths in the stories from `./Block` to `reablocks`
@@ -16,7 +16,7 @@ function replacePaths() {
     const code = fs.readFileSync(file, { encoding: 'utf-8' });
 
     const output = rewritePaths(code, path => {
-      if (EXCLUDED_PATHS.includes(path)) {
+      if (EXCLUDED_PATHS.some(excludedPath => path.startsWith(excludedPath))) {
         return path;
       }
